src/main/java/com/myinventoryapp/dataio/DataLoader.java
package com.myinventoryapp.dataio;

import com.myinventoryapp.datastorage.CustomerRepository;
import com.myinventoryapp.datastorage.SalesTransactionRepository;
import com.myinventoryapp.inventoryentities.Customer;
import com.myinventoryapp.inventoryentities.Product;
import com.myinventoryapp.util.FileUtils;
import com.myinventoryapp.datastorage.ProductRepository;
import com.myinventoryapp.inventoryentities.SalesTransaction;

import java.util.List;

public class DataLoader {

    public void loadAllData() {
        String productFilePath = FilePaths.getProductsFilePath();
        String customersFilePath = FilePaths.getCustomersFilePath();
        String transactionsFilePath = FilePaths.getTransactionsFilePath();

        loadProductsFromFile(productFilePath);
        loadCustomersFromFile(customersFilePath);
        loadTransactionsFromFile(transactionsFilePath);
    }

    void loadProductsFromFile(String productFilePath) {
        System.out.print("\nChecking the database of products : ");
        List<String> productListFromFile = FileUtils.readFromFile(productFilePath);
        for (String line : productListFromFile) {
            String[] lineData = line.split(",");
            String productName = lineData[0];
            String productId = lineData[1];
            int unitPrice = Integer.parseInt(lineData[2]);
            int quantity = Integer.parseInt(lineData[3]);
            Product product = new Product(productName, productId, unitPrice, quantity);
            ProductRepository.addProduct(product);
        }
    }

    void loadCustomersFromFile(String customersFilePath) {
        System.out.print("Checking the database of customers: ");
        List<String> customerListFromFile = FileUtils.readFromFile(customersFilePath);
        for (String line : customerListFromFile) {
            String[] rowData = line.split(",");
            String customerName = rowData[0];
            int totalPurchases = Integer.parseInt(rowData[1]);
            String customerID = rowData[2];
            Customer customer = new Customer(customerName, customerID, totalPurchases);
            CustomerRepository.addCustomer(customer);
        }
    }

    void loadTransactionsFromFile(String transactionsFilePath) {
        System.out.print("Checking the database of transactions : ");
        List<String> transactionListFromFile = FileUtils.readFromFile(transactionsFilePath);
        for (String line : transactionListFromFile) {
            String[] rowData = line.split(",");
            String transactionId = rowData[0];
            String transactionDate = rowData[1];
            String productName = rowData[2];
            int quantitySold = Integer.parseInt(rowData[3]);
            int unitPrice = Integer.parseInt(rowData[4]);
            String customerName = rowData[5];
            String customerID = rowData[6];
            SalesTransaction salesTransaction = new SalesTransaction(
                    transactionId, customerName, customerID, productName, quantitySold, unitPrice, transactionDate);
            SalesTransactionRepository.addSalesTransaction(salesTransaction);
        }
    }
}
